+++ src/main/java/choral/examples/futures/concurrentproducers/Worker2.java
@@ -43,10 +43,10 @@ public class Worker2 {
             long endTime = System.currentTimeMillis();
             System.out.println(endTime - startTime);
 
-            Iterable<Long> latencies = state.getLatencies();
+            Iterable<Float> latencies = state.getLatencies();
             try (BufferedWriter writer = Files.newBufferedWriter(Paths.get("data/concurrentproducers/worker2-latencies.csv"))) {
-                for (long value : latencies) {
-                    writer.write(Long.toString(value));
+                for (float value : latencies) {
+                    writer.write(Float.toString(value));
                     writer.newLine();
                 }
             } catch (IOException e) {
